feat: Implement transaction details dialog and integrate with MIA/NYC tabs
