+++ src/components/transaction-list.tsx
@@ -13,35 +13,16 @@ import {
   Table,
   Select,
   createListCollection,
-  useDisclosure,
-  Dialog,
-  List,
 } from "@chakra-ui/react";
 import { IoMdRefresh } from "react-icons/io";
 import { useAtom, useAtomValue, useSetAtom } from "jotai";
 import { transactionFetchStatusAtom, transactionsAtom } from "../store/stacks";
-import { formatDate, formatMicroAmount } from "../store/common";
+import { formatDate } from "../store/common";
 import { Transaction } from "@stacks/stacks-blockchain-api-types";
-import { Fragment, useState } from "react";
-import { decodeTxArgs, isValidMiningTxArgs, isValidStackingTxArgs, isValidMiningClaimTxArgs, isValidStackingClaimTxArgs } from "../utilities/transactions";
 
 interface TransactionListProps {
   transactions: Transaction[];
-}
-
-interface TransactionFunctionArgsProps {
-  functionArgs: {
-    hex: string;
-    repr: string;
-    name: string;
-    type: string;
-  }[];
-}
-
-interface TransactionDetailsDialogProps {
-  tx: Transaction | null;
-  isOpen: boolean;
-  onClose: () => void;
+  onOpenDetails: (tx: Transaction) => void;
 }
 
 function shortenPrincipal(addr: string): string {
@@ -101,7 +82,7 @@ const filterStatusCollection = createListCollection({
   ],
 });
 
-function TransactionList({ transactions }: TransactionListProps) {
+function TransactionList({ transactions, onOpenDetails }: TransactionListProps) {
   const [allTransactions, updateTransactions] = useAtom(transactionsAtom);
   const { isLoading, error, progress } = useAtomValue(
     transactionFetchStatusAtom
@@ -119,9 +100,6 @@ function TransactionList({ transactions }: TransactionListProps) {
     return true;
   });
 
-  const { open: isOpen, onOpen, onClose } = useDisclosure();
-  const [selectedTx, setSelectedTx] = useState<Transaction | null>(null);
-
   const fetchTransactions = async () => {
     if (isLoading) return;
     try {
@@ -132,8 +110,7 @@ function TransactionList({ transactions }: TransactionListProps) {
   };
 
   const handleOpenDetails = (tx: Transaction) => {
-    setSelectedTx(tx);
-    onOpen();
+    onOpenDetails(tx);
   };
 
   // Compute summaries
@@ -151,305 +128,164 @@ function TransactionList({ transactions }: TransactionListProps) {
   );
 
   return (
-    <>
-      <Stack gap={4}>
-        <Stack direction="row" alignItems="center" minH="2em">
-          <Box
-            w={3}
-            h={3}
-            borderRadius="50%"
-            bg={isLoading ? "yellow.500" : error ? "red.500" : "green.500"}
-          />
-          {isLoading && (
-            <Stack
-              direction="row"
-              align="center"
-              justifyContent="space-between"
-              w="100%"
-            >
-              <Text>Loading transactions... {progress}%</Text>
-              <Spinner size="sm" />
-            </Stack>
-          )}
-
-          {error && <Text color="red.500">Error: {error}</Text>}
-
-          {!isLoading && !error && (
-            <Stack
-              direction="row"
-              align="center"
-              justifyContent="space-between"
-              w="100%"
-            >
-              <Text>
-                {allTransactions.length > 0
-                  ? `Filtered transactions: ${filteredTransactions.length} / ${allTransactions.length}`
-                  : "No transactions loaded yet"}
-              </Text>
-              <IconButton
-                aria-label="Refresh Transactions"
-                title="Refresh Transactions"
-                size="sm"
-                onClick={fetchTransactions}
-              >
-                <IoMdRefresh />
-              </IconButton>
-            </Stack>
-          )}
-        </Stack>
-        <Stack direction="row" gap={4} flexWrap="wrap">
-          <Badge colorScheme="green" variant="outline">Mining: {summaries.mining}</Badge>
-          <Badge colorScheme="blue" variant="outline">Mining Claims: {summaries.miningClaims}</Badge>
-          <Badge colorScheme="purple" variant="outline">Stacking: {summaries.stacking}</Badge>
-          <Badge colorScheme="orange" variant="outline">Stacking Claims: {summaries.stackingClaims}</Badge>
-          <Badge colorScheme="yellow" variant="outline">Transfers: {summaries.transfers}</Badge>
-        </Stack>
-        <Stack direction="row" gap={4} flexWrap="wrap">
-          <Select.Root
-            collection={filterTypeCollection}
-            value={[filterType]}
-            onValueChange={(e) => setFilterType(e.value[0])}
-            size="sm"
-            w="auto"
+    <Stack gap={4}>
+      <Stack direction="row" alignItems="center" minH="2em">
+        <Box
+          w={3}
+          h={3}
+          borderRadius="50%"
+          bg={isLoading ? "yellow.500" : error ? "red.500" : "green.500"}
+        />
+        {isLoading && (
+          <Stack
+            direction="row"
+            align="center"
+            justifyContent="space-between"
+            w="100%"
           >
-            <Select.HiddenSelect />
-            <Select.Control>
-              <Select.Trigger>
-                <Select.ValueText>All Types</Select.ValueText>
-              </Select.Trigger>
-              <Select.IndicatorGroup>
-                <Select.Indicator />
-              </Select.IndicatorGroup>
-            </Select.Control>
-            <Portal>
-              <Select.Positioner>
-                <Select.Content>
-                  {filterTypeCollection.items.map((item) => (
-                    <Select.Item item={item} key={item.value}>
-                      {item.label}
-                      <Select.ItemIndicator />
-                    </Select.Item>
-                  ))}
-                </Select.Content>
-              </Select.Positioner>
-            </Portal>
-          </Select.Root>
-          <Select.Root
-            collection={filterStatusCollection}
-            value={[filterStatus]}
-            onValueChange={(e) => setFilterStatus(e.value[0])}
-            size="sm"
-            w="auto"
-          >
-            <Select.HiddenSelect />
-            <Select.Control>
-              <Select.Trigger>
-                <Select.ValueText>All Statuses</Select.ValueText>
-              </Select.Trigger>
-              <Select.IndicatorGroup>
-                <Select.Indicator />
-              </Select.IndicatorGroup>
-            </Select.Control>
-            <Portal>
-              <Select.Positioner>
-                <Select.Content>
-                  {filterStatusCollection.items.map((item) => (
-                    <Select.Item item={item} key={item.value}>
-                      {item.label}
-                      <Select.ItemIndicator />
-                    </Select.Item>
-                  ))}
-                </Select.Content>
-              </Select.Positioner>
-            </Portal>
-          </Select.Root>
-          <Input placeholder="Search by TXID" value={searchTerm} onChange={e => setSearchTerm(e.target.value)} w="auto" />
-        </Stack>
-        <Box overflowX="auto">
-          {filteredTransactions.length === 0 && <Text>No transactions found.</Text>}
-          {filteredTransactions.length > 0 && (
-            <Table.Root variant="outline">
-              <Table.Header>
-                <Table.Row>
-                  <Table.ColumnHeader>TXID</Table.ColumnHeader>
-                  <Table.ColumnHeader>Type</Table.ColumnHeader>
-                  <Table.ColumnHeader>Status</Table.ColumnHeader>
-                  <Table.ColumnHeader>Date</Table.ColumnHeader>
-                  <Table.ColumnHeader>Actions</Table.ColumnHeader>
-                </Table.Row>
-              </Table.Header>
-              <Table.Body>
-                {filteredTransactions.map((tx) => {
-                  const category = getCategory(tx);
-                  return (
-                    <Table.Row key={tx.tx_id}>
-                      <Table.Cell>
-                        <Link href={`https://explorer.hiro.so/tx/${tx.tx_id}`} rel="noopener noreferrer" target="_blank">
-                          {shortenTxId(tx.tx_id)}
-                        </Link>
-                      </Table.Cell>
-                      <Table.Cell>
-                        <Badge colorScheme={getCategoryColor(category)}>{category}</Badge>
-                      </Table.Cell>
-                      <Table.Cell>
-                        <Badge colorScheme={tx.tx_status === 'success' ? 'green' : 'red'}>
-                          {tx.tx_status}
-                        </Badge>
-                      </Table.Cell>
-                      <Table.Cell>{formatDate(tx.block_time_iso)}</Table.Cell>
-                      <Table.Cell>
-                        <Button size="sm" onClick={() => handleOpenDetails(tx)}>
-                          Details
-                        </Button>
-                      </Table.Cell>
-                    </Table.Row>
-                  );
-                })}
-              </Table.Body>
-            </Table.Root>
-          )}
-        </Box>
-      </Stack>
-      <TransactionDetailsDialog
-        tx={selectedTx}
-        isOpen={isOpen}
-        onClose={onClose}
-      />
-    </>
-  );
-}
-
-
-function TransactionFunctionArgs({
-  functionArgs,
-}: TransactionFunctionArgsProps) {
-  return (
-    <Stack>
-      <Text fontWeight="bold">Function Arguments</Text>
-      <List.Root gap={2}>
-        {functionArgs.map((arg) => (
-          <List.Item key={arg.hex}>
-            <Text>Name: {arg.name}</Text>
-            <Text>Type: {arg.type}</Text>
-            <Text>Repr: {arg.repr}</Text>
-          </List.Item>
-        ))}
-      </List.Root>
-    </Stack>
-  );
-}
-
-function DecodedFunctionArgs({ tx }: { tx: Transaction }) {
-  let decoded;
-  try {
-    decoded = decodeTxArgs(tx);
-  } catch (error) {
-    return <Text>Failed to decode arguments: {error instanceof Error ? error.toString() : `Unknown error: ${String(error)}`}</Text>;
-  }
-
-  if (!decoded) {
-    return <Text>No decodable arguments.</Text>;
-  }
-
-  let decodedType = 'Unknown';
-  let gridItems: { label: string; value: string }[] = [];
-
-  if (isValidMiningTxArgs(decoded)) {
-    decodedType = 'Mining';
-    gridItems = [
-      { label: 'Amounts uSTX', value: decoded.amountsUstx.map(a => a.toString()).join(', ') },
-    ];
-  } else if (isValidStackingTxArgs(decoded)) {
-    decodedType = 'Stacking';
-    gridItems = [
-      { label: 'Amount Token', value: decoded.amountToken.toString() },
-      { label: 'Lock Period', value: decoded.lockPeriod.toString() },
-    ];
-  } else if (isValidMiningClaimTxArgs(decoded)) {
-    decodedType = 'Mining Claim';
-    gridItems = [
-      { label: 'Miner Block Height', value: decoded.minerBlockHeight.toString() },
-    ];
-  } else if (isValidStackingClaimTxArgs(decoded)) {
-    decodedType = 'Stacking Claim';
-    gridItems = [
-      { label: 'Reward Cycle', value: decoded.rewardCycle.toString() },
-    ];
-  }
-
-  return (
-    <Stack>
-      <Text fontWeight="bold">Decoded Arguments ({decodedType})</Text>
-      <Grid templateColumns="1fr 1fr" gap={2}>
-        {gridItems.map((item, index) => (
-          <Fragment key={index}>
-            <Text>{item.label}:</Text>
-            <Text>{item.value}</Text>
-          </Fragment>
-        ))}
-      </Grid>
-    </Stack>
-  );
-}
-
-function TransactionDetailsDialog({
-  tx,
-  isOpen,
-  onClose,
-}: TransactionDetailsDialogProps) {
-  if (!tx) return null;
-
-  return (
-    <Dialog.Root open={isOpen} onOpenChange={onClose} size="full">
-      <Dialog.Backdrop />
-      <Dialog.Content>
-        <Dialog.Header>
-          <Dialog.Title>Transaction Details</Dialog.Title>
-          <Dialog.CloseTrigger />
-        </Dialog.Header>
-        <Dialog.Body>
-          <Stack gap={4}>
-            <Grid templateColumns="1fr 3fr" gap={2}>
-              <Text fontWeight="bold">TXID:</Text>
-              <Link href={`https://explorer.hiro.so/tx/${tx.tx_id}`} rel="noopener noreferrer" target="_blank">
-                {tx.tx_id}
-              </Link>
-              <Text fontWeight="bold">Status:</Text>
-              <Badge colorScheme={tx.tx_status === 'success' ? 'green' : 'red'}>
-                {tx.tx_status}
-              </Badge>
-              <Text fontWeight="bold">Block Height:</Text>
-              <Link href={`https://explorer.hiro.so/block/${tx.block_height}`} rel="noopener noreferrer" target="_blank">
-                {tx.block_height}
-              </Link>
-              <Text fontWeight="bold">Block Time:</Text>
-              <Text>{formatDate(tx.block_time_iso)}</Text>
-              <Text fontWeight="bold">Sender Address:</Text>
-              <Text>{shortenPrincipal(tx.sender_address)}</Text>
-              <Text fontWeight="bold">Fee:</Text>
-              <Text>{formatMicroAmount(parseFloat(tx.fee_rate))} STX</Text>
-            </Grid>
-            {tx.tx_type === "contract_call" && (
-              <Stack gap={2}>
-                <Text fontWeight="bold" fontSize="lg">Contract Call Details</Text>
-                <Grid templateColumns="1fr 3fr" gap={2}>
-                  <Text fontWeight="bold">Contract ID:</Text>
-                  <Text>{shortenPrincipal(tx.contract_call.contract_id)}</Text>
-                  <Text fontWeight="bold">Function Name:</Text>
-                  <Text>{tx.contract_call.function_name}</Text>
-                </Grid>
-                {tx.contract_call.function_args && (
-                  <TransactionFunctionArgs
-                    functionArgs={tx.contract_call.function_args}
-                  />
-                )}
-                <DecodedFunctionArgs tx={tx} />
-              </Stack>
-            )}
+            <Text>Loading transactions... {progress}%</Text>
+            <Spinner size="sm" />
           </Stack>
-        </Dialog.Body>
-      </Dialog.Content>
-    </Dialog.Root>
+        )}
+
+        {error && <Text color="red.500">Error: {error}</Text>}
+
+        {!isLoading && !error && (
+          <Stack
+            direction="row"
+            align="center"
+            justifyContent="space-between"
+            w="100%"
+          >
+            <Text>
+              {allTransactions.length > 0
+                ? `Filtered transactions: ${filteredTransactions.length} / ${allTransactions.length}`
+                : "No transactions loaded yet"}
+            </Text>
+            <IconButton
+              aria-label="Refresh Transactions"
+              title="Refresh Transactions"
+              size="sm"
+              onClick={fetchTransactions}
+            >
+              <IoMdRefresh />
+            </IconButton>
+          </Stack>
+        )}
+      </Stack>
+      <Stack direction="row" gap={4} flexWrap="wrap">
+        <Badge colorScheme="green" variant="outline">Mining: {summaries.mining}</Badge>
+        <Badge colorScheme="blue" variant="outline">Mining Claims: {summaries.miningClaims}</Badge>
+        <Badge colorScheme="purple" variant="outline">Stacking: {summaries.stacking}</Badge>
+        <Badge colorScheme="orange" variant="outline">Stacking Claims: {summaries.stackingClaims}</Badge>
+        <Badge colorScheme="yellow" variant="outline">Transfers: {summaries.transfers}</Badge>
+      </Stack>
+      <Stack direction="row" gap={4} flexWrap="wrap">
+        <Select.Root
+          collection={filterTypeCollection}
+          value={[filterType]}
+          onValueChange={(e) => setFilterType(e.value[0])}
+          size="sm"
+          w="auto"
+        >
+          <Select.HiddenSelect />
+          <Select.Control>
+            <Select.Trigger>
+              <Select.ValueText>All Types</Select.ValueText>
+            </Select.Trigger>
+            <Select.IndicatorGroup>
+              <Select.Indicator />
+            </Select.IndicatorGroup>
+          </Select.Control>
+          <Portal>
+            <Select.Positioner>
+              <Select.Content>
+                {filterTypeCollection.items.map((item) => (
+                  <Select.Item item={item} key={item.value}>
+                    {item.label}
+                    <Select.ItemIndicator />
+                  </Select.Item>
+                ))}
+              </Select.Content>
+            </Select.Positioner>
+          </Portal>
+        </Select.Root>
+        <Select.Root
+          collection={filterStatusCollection}
+          value={[filterStatus]}
+          onValueChange={(e) => setFilterStatus(e.value[0])}
+          size="sm"
+          w="auto"
+        >
+          <Select.HiddenSelect />
+          <Select.Control>
+            <Select.Trigger>
+              <Select.ValueText>All Statuses</Select.ValueText>
+            </Select.Trigger>
+            <Select.IndicatorGroup>
+              <Select.Indicator />
+            </Select.IndicatorGroup>
+          </Select.Control>
+          <Portal>
+            <Select.Positioner>
+              <Select.Content>
+                {filterStatusCollection.items.map((item) => (
+                  <Select.Item item={item} key={item.value}>
+                    {item.label}
+                    <Select.ItemIndicator />
+                  </Select.Item>
+                ))}
+              </Select.Content>
+            </Select.Positioner>
+          </Portal>
+        </Select.Root>
+        <Input placeholder="Search by TXID" value={searchTerm} onChange={e => setSearchTerm(e.target.value)} w="auto" />
+      </Stack>
+      <Box overflowX="auto">
+        {filteredTransactions.length === 0 && <Text>No transactions found.</Text>}
+        {filteredTransactions.length > 0 && (
+          <Table.Root variant="outline">
+            <Table.Header>
+              <Table.Row>
+                <Table.ColumnHeader>TXID</Table.ColumnHeader>
+                <Table.ColumnHeader>Type</Table.ColumnHeader>
+                <Table.ColumnHeader>Status</Table.ColumnHeader>
+                <Table.ColumnHeader>Date</Table.ColumnHeader>
+                <Table.ColumnHeader>Actions</Table.ColumnHeader>
+              </Table.Row>
+            </Table.Header>
+            <Table.Body>
+              {filteredTransactions.map((tx) => {
+                const category = getCategory(tx);
+                return (
+                  <Table.Row key={tx.tx_id}>
+                    <Table.Cell>
+                      <Link href={`https://explorer.hiro.so/tx/${tx.tx_id}`} rel="noopener noreferrer" target="_blank">
+                        {shortenTxId(tx.tx_id)}
+                      </Link>
+                    </Table.Cell>
+                    <Table.Cell>
+                      <Badge colorScheme={getCategoryColor(category)}>{category}</Badge>
+                    </Table.Cell>
+                    <Table.Cell>
+                      <Badge colorScheme={tx.tx_status === 'success' ? 'green' : 'red'}>
+                        {tx.tx_status}
+                      </Badge>
+                    </Table.Cell>
+                    <Table.Cell>{formatDate(tx.block_time_iso)}</Table.Cell>
+                    <Table.Cell>
+                      <Button size="sm" onClick={() => handleOpenDetails(tx)}>
+                        Details
+                      </Button>
+                    </Table.Cell>
+                  </Table.Row>
+                );
+              })}
+            </Table.Body>
+          </Table.Root>
+        )}
+      </Box>
+    </Stack>
   );
 }
